praktika/requirements.txt
pre-commit==3.0.1
requests==2.31.0
python-telegram-bot==21.3
psycopg2==2.9.9